#61 - Implement Mac Touch Bar Support! - created shared-interface, removed duplicates
