+++ interfaces/shared-interfaces.ts
@@ -1,0 +1,16 @@
+// Let's make these identical to settings buttons!
+export type SupportedView = 'showThumbnails'
+  | 'showFilmstrip'
+  | 'showFullView'
+  | 'showDetails'
+  | 'showFiles'
+  | 'showClips';
+
+export const allSupportedViews: SupportedView[] = [
+  'showThumbnails',
+  'showFilmstrip',
+  'showFullView',
+  'showDetails',
+  'showFiles',
+  'showClips',
+];
